Migrate utils to TypeScript
