+++ app/externalModules.d.ts
@@ -14,3 +14,4 @@ declare module 'react-native-restart';
 declare module 'react-native-prompt-android';
 declare module 'react-native-jitsi-meet';
 declare module 'rn-root-view';
+declare module 'semver';
